Address CodeRabbit review on #34: correctness + SSRF dedup
- app/routes.py: re-raise HTTPException in the image-gen handler so a
  deliberate 400 (missing URL) isn't collapsed into a 500; return 4xx (not
  503) when a caller-supplied llm_provider fails to resolve.
- app/models.py: align speech_speed bounds to the runner's honored 0.6–1.5
  (it clamps to that range, so wider bounds silently lied); fix the copy-paste
  MCPConfig docstring.
- README.md: "four" OpenAPI snapshots, not three.
- SSRF hardening was duplicated verbatim across utils/mcp_client.py and
  app/services/prompt_context.py. Centralize the shared primitives
  (PinnedResolver, is_private_ip, build_pinned_connector) in new utils/ssrf.py;
  each caller keeps its own validation wrapper (exception type, status code,
  allow-private bypass policy differ). All 18 mcp/prompt tests still pass.
