+++ utils/mcp_client.py
@@ -10,9 +10,10 @@ import re
 import socket
 from collections.abc import Mapping, Sequence
 from dataclasses import dataclass, field
-from ipaddress import ip_address
 from typing import Any
 from urllib.parse import urlparse
+
+from utils.ssrf import build_pinned_connector, is_private_ip
 
 
 LOGGER = logging.getLogger(__name__)
@@ -30,37 +31,6 @@ SECRET_KEY_PARTS = (
     "token",
 )
 aiohttp: Any | None = None
-
-
-class _PinnedResolver:
-    """aiohttp resolver that only returns pre-validated addresses for one host."""
-
-    def __init__(self, addresses_by_host: dict[str, list[str]]):
-        self._addresses_by_host = addresses_by_host
-
-    async def resolve(
-        self,
-        host: str,
-        port: int = 0,
-        family: int = socket.AF_INET,
-    ) -> list[dict[str, Any]]:
-        addresses = self._addresses_by_host.get(host)
-        if not addresses:
-            raise OSError(f"Host {host} was not pre-validated")
-        return [
-            {
-                "hostname": host,
-                "host": address,
-                "port": port,
-                "family": socket.AF_INET6 if ":" in address else socket.AF_INET,
-                "proto": 0,
-                "flags": socket.AI_NUMERICHOST,
-            }
-            for address in addresses
-        ]
-
-    async def close(self) -> None:
-        return None
 
 
 class McpClientError(Exception):
@@ -108,18 +78,6 @@ def _private_mcp_url_bypass_enabled(url: str) -> bool:
     return _private_mcp_urls_allowed() or _is_allowed_private_mcp_url(url)
 
 
-def _is_private_ip(value: str) -> bool:
-    parsed = ip_address(value)
-    return (
-        parsed.is_private
-        or parsed.is_loopback
-        or parsed.is_link_local
-        or parsed.is_multicast
-        or parsed.is_reserved
-        or parsed.is_unspecified
-    )
-
-
 async def validate_mcp_http_url(url: str) -> list[str] | None:
     """Block SSRF targets and return DNS answers pinned into aiohttp."""
     parsed = urlparse(url)
@@ -131,7 +89,7 @@ async def validate_mcp_http_url(url: str) -> list[str] | None:
 
     host = parsed.hostname
     try:
-        if _is_private_ip(host):
+        if is_private_ip(host):
             raise McpClientError(f"MCP HTTP URL host is private or local: {host}")
         return [host]
     except ValueError:
@@ -147,7 +105,7 @@ async def validate_mcp_http_url(url: str) -> list[str] | None:
     resolved_ips = []
     for address in addresses:
         resolved_ip = address[4][0]
-        if _is_private_ip(resolved_ip):
+        if is_private_ip(resolved_ip):
             raise McpClientError(
                 f"MCP HTTP URL resolves to private or local address: {host}"
             )
@@ -156,20 +114,9 @@ async def validate_mcp_http_url(url: str) -> list[str] | None:
     return resolved_ips
 
 
-def build_pinned_mcp_connector(
-    aiohttp_module: Any,
-    url: str,
-    resolved_ips: list[str] | None,
-) -> Any | None:
-    if not resolved_ips or not hasattr(aiohttp_module, "TCPConnector"):
-        return None
-    parsed = urlparse(url)
-    if not parsed.hostname:
-        return None
-    return aiohttp_module.TCPConnector(
-        resolver=_PinnedResolver({parsed.hostname: resolved_ips}),
-        ttl_dns_cache=0,
-    )
+# Shared SSRF connector builder (see utils/ssrf.py). Kept under this name for
+# the call site below.
+build_pinned_mcp_connector = build_pinned_connector
 
 
 def split_mcp_runtime_headers(mcp_config: Any) -> tuple[dict[str, Any], list[dict[str, str] | None]]:
